Replaced 02-dashboard-ui with 02-onboarding, 03-plugin-settings
