--- tests/qa/tests/02-dashboard-ui/_test-data/index.ts
@@ -1,5 +1,0 @@
-export * from './learn-more-links.data';
-export * from './pay-later-messaging.data';
-export * from './payment-methods.data';
-export * from './ui-tests-per-country.data';
-export * from './badges.data';
