+++ tests/qa/tests/02-onboarding/_test-data/index.ts
@@ -1,0 +1,3 @@
+export * from './onboarding-learn-more-links.data';
+export * from './onboarding-ui.data';
+export * from './onboarding-badges.data';
