+++ tests/qa/tests/02-onboarding/_test-data/onboarding-badges.data.ts
