+++ tests/qa/tests/02-onboarding/_test-data/onboarding-ui.data.ts
