+++ tests/qa/tests/02-onboarding/onboarding-badges.spec.ts
@@ -3,7 +3,6 @@
  */
 import {
 	test,
-	expect,
 	saveTestResultsToFile,
 	getTestResultsFromFile,
 } from '../../utils';
