+++ tests/qa/tests/02-onboarding/onboarding-subscription.spec.ts
@@ -2,7 +2,12 @@
  * Internal dependencies
  */
 import { test, expect } from '../../utils';
-import { storeConfigDefault, subscriptionsPlugin, merchants, Pcp } from '../../resources';
+import {
+	storeConfigDefault,
+	subscriptionsPlugin,
+	merchants,
+	Pcp,
+} from '../../resources';
 
 test.beforeAll( async ( { utils, pcpApi } ) => {
 	await utils.configureStore( storeConfigDefault );
@@ -47,7 +52,7 @@ test.describe( () => {
 		await requestUtils.deactivatePlugin( subscriptionsPlugin.slug );
 	} );
 
-	test( 'PCP-4357 | Subscription - Settings - US - Onboarding - Connect with business account, all product types, card payment enabled', async ( {
+	test.only( 'PCP-4357 | Subscription - Settings - US - Onboarding - Connect with business account, all product types, card payment enabled', async ( {
 		pcpOnboarding,
 		pcpPaymentMethods,
 		pcpSettings,
@@ -55,42 +60,44 @@ test.describe( () => {
 		pcpApi,
 	}, testInfo ) => {
 		await pcpOnboarding.visit();
-	
+
 		await pcpOnboarding.activatePayPalPaymentsButton().click();
-	
+
 		await pcpOnboarding.businessRadio().click();
 		await pcpOnboarding.continueButton().click();
-	
+
 		await pcpOnboarding.physicalGoodsCheckbox().check();
 		await pcpOnboarding.subscriptionsCheckbox().check();
 		await pcpOnboarding.continueButton().click();
-	
+
 		await pcpOnboarding.chooseOptionalPaymentMethods( true );
 		await pcpOnboarding.continueButton().click();
 
 		await pcpOnboarding.gotoInitialOnboardingPage();
-	
+
 		await pcpApi.connectMerchant(
 			merchants.usa.client_id,
 			merchants.usa.client_secret
 		);
-	
+
 		await pcpOnboarding.page.reload();
 		await pcpOnboarding.snapshotContent(
-			`${ testInfo.title } - Overview`, 3000
+			`${ testInfo.title } - Overview`,
+			3000
 		);
-	
-	
+
 		await pcpSettings.visit();
 		await pcpOnboarding.snapshotContent(
-			`${ testInfo.title } - Settings - Pay Now enabled by default`, 3000
+			`${ testInfo.title } - Settings - Pay Now enabled by default`,
+			3000
 		);
-	
+
 		await pcpPaymentMethods.visit();
 		await pcpOnboarding.snapshotContent(
-			`${ testInfo.title } - Payment methods - PayPal, Venmo enabled`, 3000
+			`${ testInfo.title } - Payment methods - PayPal, Venmo enabled`,
+			3000
 		);
-	
+
 		const snapshotName = testInfo.title;
 		const locations: Pcp.Admin.Styling.Location[] = [
 			'Cart',
@@ -99,11 +106,11 @@ test.describe( () => {
 			'Mini Cart',
 			'Product Page',
 		];
-	
+
 		await pcpStyling.visit();
 		await expect( pcpStyling.configContainer() ).toBeVisible();
 		await expect( pcpStyling.locationSelectbox() ).toBeVisible();
-	
+
 		for ( const location of locations ) {
 			await pcpStyling.locationSelectbox().selectOption( location );
 			await pcpStyling.snapshotStylingConfigurator(
