+++ tests/qa/tests/03-plugin-settings/_test-data/index.ts
@@ -1,0 +1,2 @@
+export * from './pay-later-messaging.data';
+export * from './payment-methods-ui.data';
