+++ tests/qa/tests/03-plugin-settings/payment-methods-ui.spec.ts
